fix: db popup inconsistent position
