+++ lib/static/components/header/summary/dbSummary.js
@@ -1,8 +1,8 @@
 import React, {Component} from 'react';
-import {Dropdown, Button, Ref} from 'semantic-ui-react';
+import {Dropdown} from 'semantic-ui-react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
 
+import DbBtn from './dbBtn';
 import Popup from '../../popup';
 
 export default class DbSummaryKey extends Component {
@@ -28,9 +28,6 @@ export default class DbSummaryKey extends Component {
             return null;
         }
 
-        const successFetchDbDetails = fetchDbDetails.filter(d => d.success);
-        const isFailed = successFetchDbDetails.length !== fetchDbDetails.length;
-        const value = `${successFetchDbDetails.length}/${fetchDbDetails.length}`;
         const additionalInfo = fetchDbDetails.map(({url, status, success}) => (
             <Dropdown.Item key={url} className="db-info__row">
                 {' '}
@@ -38,28 +35,12 @@ export default class DbSummaryKey extends Component {
                 <span className={success ? 'db-info__row_success' : 'db-info__row_fail'}>{' ' + status}</span>
             </Dropdown.Item>
         ));
-        const title = `Databases loaded: ${value}`;
-        const btnClassNames = classNames(
-            'db-info',
-            {'db-info_failed': isFailed}
-        );
-
-        const DbBtn = React.forwardRef((props, ref) => (
-            <Ref innerRef={ref}>
-                <Button
-                    content={title}
-                    icon="angle down"
-                    className={btnClassNames}
-                    basic
-                />
-            </Ref>
-        ));
 
         return (
             <Popup
                 className="db-info__popup"
                 action="hover"
-                target={<DbBtn />}
+                target={<DbBtn fetchDbDetails={fetchDbDetails} />}
             >
                 {additionalInfo}
             </Popup>
